Fixed precision problem and removed 'Z' from isoDateTime
'Z' was inappropriate since no timezone info was being given to the
the function. The precision problem was causing the ostream operator
to round the number up when the seconds value was truncated to
millisecond precision. This occasionally resulted in an illegal seconds
value of '60.000'. Using nanosecond precision prevents this.